Completed phong
I completed the phong shading within the vertex shader. Next step is putting it in the fragment for more accurate shading.

Found that the power needs to be below 1.0f for the specular to show correctly. Need to ask Marius to look at my code and explain why because I can create the anti-specular but I do not understand this power.