Completed Surface animation
Tried doing the surface animation in after passing initial vertex data to textures. Just ended up multiplying with 0 so nothing would render. Have decided to animate and then defer for a lighting pass.

This means it will likely be quite difficult to animate using the NoiseMap within the existing framework...