Improved vertex animation and debugging noise map
Added movement to the z-values of position so that the waves look a bit more 3D.

Got the noise texture generation working, but struggling to apply it to only the plane. Believe this is because all objects are in the same, default VAO, but because of how the framebuffer/VAO's structured I am getting a constant, repeating error.

 Next aim is to create new VAO and attempt to render the background and the plane separately. So that I can use separate textures on them. I may need to review how textures work...Or how I did it in the first coursework.